refactoring and added test case for rule engine

+++ src/main/java/com/vehicle/rental/app/service/ITripReservationService.java
@@ -4,6 +4,6 @@ import com.vehicle.rental.app.domain.Trip;
 
 public interface ITripReservationService {
 
-	public Integer calculateTripExpense(Trip trip);
+	public Double calculateTripExpense(Trip trip);
 	
 }
